model_train/deepface_algorithm/requirements.txt
numpy<2.0
opencv-python
pillow
scikit-learn
matplotlib
requests
deepface
pandas
tensorflow
